refactor: move lógica de autenticação e login para pasta auth
