+++ internal/auth/auth.go
@@ -1,4 +1,4 @@
-package model
+package auth
 
 type LoginRequest struct {
 	Email    string `json:"email" validate:"required,email"`
