+++ internal/auth/auth_service.go
@@ -1,4 +1,4 @@
-package service
+package auth
 
 import (
 	"errors"
